MAK-1 Initial branding updates.
Modified the Ghost Jekyll theme from jekyller/jasper2 with Makánal branding and basic welcome message. To work with GitHub personal pages, the master branch was renamed to dev-master. When new commits are added. The contents of the _site directory must be copied to the master branch and committed. GitHub will deploy the static site from the root of the master branch.