Merged PR 1: TASK-1 Adding _site directory back to master branch.
TASK-1 Adding _site directory back to master branch.

The Azure Web App is configured to load the static site from the _site directory which was not possible on GitHub. Now that the site will be deployed on Azure, there's no need for the separate master and dev-master branches.

Related work items: #1