TASK-3 Replaced all active http links with https or //.
Without this update, we're unable to release as https due to pages containing insecure links. This also prevents us from releasing to CDN with https. All links on active pages have been updated to either https in "always secure" cases or // if dependent on how the page was loaded (http vs https).