TASK-27 Added Client Services Page.
Create new Services page and added link to navigation bar. Additional updates were made to remove photos from original design, optimize site photos for page speed load improvements and repository size management. The Welcome post was also updated with a friendlier message noting it's a work in progress and will be completed soon. The Welcome post now directs the reader to the new Services page since it currently has the most complete content.

This update also introduces new CSS stylings. In global.css the `<small>` tag was reduced to 50% max-width, which makes it useful for photo credit lines. For now, the original selector for the `<small>` tag is commented out and may be removed in a future release. In screen.css, `#left` and `#right` can now be used (similar to `#full`) to align images to the left or right in posts. The max-width is set to 50% which gives a nice alternating or column-like layout.

Signed-off-by: Kane McConnell <kane@makanal.eu>
CLA-Signature: 62cbcb4130cbddfe82c60e39d0853db8